+++ Source/Private/New-SuggestedCorrection.ps1
@@ -12,7 +12,7 @@
 #>
 function New-SuggestedCorrection
 {
-    [CmdletBinding()]
+    [CmdletBinding(SupportsShouldProcess = $true,ConfirmImpact='None')]
     [OutputType([Microsoft.Windows.PowerShell.ScriptAnalyzer.Generic.CorrectionExtent])]
     param (
         [Parameter(Mandatory = $true)]
@@ -28,13 +28,17 @@ function New-SuggestedCorrection
         $Description
     )
 
-    [Microsoft.Windows.PowerShell.ScriptAnalyzer.Generic.CorrectionExtent]::new(
-        $Extent.StartLineNumber,
-        $Extent.EndLineNumber,
-        $Extent.StartColumnNumber,
-        $Extent.EndColumnNumber,
-        $NewString,
-        $Extent.File,
-        $Description
-    )
+    if ($PSCmdlet.ShouldProcess("Create correction extent"))
+    {
+        [Microsoft.Windows.PowerShell.ScriptAnalyzer.Generic.CorrectionExtent]::new(
+            $Extent.StartLineNumber,
+            $Extent.EndLineNumber,
+            $Extent.StartColumnNumber,
+            $Extent.EndColumnNumber,
+            $NewString,
+            $Extent.File,
+            $Description
+        )
+    }
+
 }